+++ src/__swaps__/utils/meteorology.ts
